lift: Share common lifter arguments.
