+++ bin/lift/Lift.cpp
@@ -69,30 +69,6 @@ DEFINE_string(lift_with, "", "");
 
 namespace cli = circ::cli;
 
-namespace circ::cli
-{
-    struct ConjureALU : circ::DefaultCmdOpt, Arity< 0 >
-    {
-        static inline const auto opt = circ::CmdOpt( "--conjure-alu", false );
-    };
-
-    struct NoAdvices : circ::DefaultCmdOpt, Arity< 0 >
-    {
-        static inline const auto opt = circ::CmdOpt( "--no-advices", false );
-    };
-
-    struct LiftWith : DefaultCmdOpt, HasAllowed< LiftWith >,
-                      PathArg
-    {
-        static inline const auto opt = circ::CmdOpt( "--lift-with", true );
-
-        static inline const std::unordered_set< std::string > allowed =
-        {
-            "mux-heavy", "disjunctions", "v3"
-        };
-    };
-};
-
 using circuit_owner_t = circ::circuit_owner_t;
 
 namespace
